change fixnumber to float

+++ lesson1/task2.rb
@@ -1,6 +1,6 @@
 puts 'Введиту а - основание треугольника'
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 puts 'Введиту h - высота треугольника'
-h = gets.chomp.to_i
+h = gets.chomp.to_f
 s = a*h*0.5
 puts "Площадь треугольника #{s}"
